+++ vmck/backends/qemu.py
@@ -1,4 +1,3 @@
-import random
 from urllib.parse import urljoin
 from pathlib import Path
 
@@ -6,41 +5,15 @@ from django.conf import settings
 
 from vmck.backends import submission
 from vmck.backends import socat
+from vmck.backends import qemu_utils
 
 
 control_path = (Path(__file__).parent / 'control').resolve()
 
 
-def random_port(start=10000, end=20000):
-    return random.SystemRandom().randint(start, end - 1)
-
-
-def constraints():
-    return [
-        {
-            'LTarget': "${meta.vmck_worker}",
-            'RTarget': "",
-            'Operand': "is_set",
-        },
-    ]
-
-
-def resources(vm_port, options):
-    network = {
-        'ReservedPorts': [
-            {'Label': 'ssh', 'Value': vm_port},
-        ],
-    }
-    return {
-        'Networks': [network],
-        'MemoryMB': options['memory'],
-        'CPU': options['cpu_mhz'],
-    }
-
-
 def task_group(job, options):
     tasks = []
-    vm_port = random_port(
+    vm_port = qemu_utils.random_port(
         settings.VM_PORT_RANGE_START,
         settings.VM_PORT_RANGE_STOP,
     )
@@ -96,8 +69,10 @@ def task_group(job, options):
             'accelerator': 'kvm',
             'args': qemu_args,
         },
-        'resources': resources(vm_port, options),
+        'resources': qemu_utils.resources(vm_port, options),
+        'services': socat.services(job),
     }
+
     tasks.append(vm_task)
     tasks.append(socat.task(job))
 
@@ -106,7 +81,7 @@ def task_group(job, options):
 
     return {
         'name': 'test',
-        'Constraints': constraints(),
+        'Constraints': qemu_utils.constraints(),
         'tasks': tasks,
         'RestartPolicy': {
             'Attempts': 0,
